refactor(db): extract custom commands to src/db/customCommands.ts
Move all custom command logic (types, lookup cache, cache build helpers,
CRUD and user-assignment operations) into their own module. db.ts retains
thin re-export lines only; all 8 callers are unaffected.
